+++ smarts/benchmark/driving_smarts/v2023/metric_formula_drive.py
@@ -20,20 +20,19 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-import functools
 from typing import Dict
 
 import numpy as np
 
 from smarts.env.gymnasium.wrappers.metric.costs import Costs
-from smarts.env.gymnasium.wrappers.metric.formula import FormulaBase, Score
-from smarts.env.gymnasium.wrappers.metric.params import Comfort, DistToObstacles, Params
-from smarts.env.gymnasium.wrappers.metric.types import Record
-from smarts.env.gymnasium.wrappers.metric.utils import (
-    add_dataclass,
-    divide,
-    op_dataclass,
+from smarts.env.gymnasium.wrappers.metric.formula import FormulaBase, Score, avg_costs
+from smarts.env.gymnasium.wrappers.metric.params import (
+    Comfort,
+    DistToObstacles,
+    JerkLinear,
+    Params,
 )
+from smarts.env.gymnasium.wrappers.metric.types import Record
 
 
 class Formula(FormulaBase):
@@ -57,6 +56,7 @@ class Formula(FormulaBase):
             dist_to_obstacles=DistToObstacles(
                 active=False,
             ),
+            jerk_linear=JerkLinear(active=False),
         )
         return params
 
@@ -84,27 +84,7 @@ class Formula(FormulaBase):
             "HumannessError", and "RuleViolation" scores.
         """
 
-        costs_total = Costs()
-        episodes = 0
-        for scen, val in records_sum.items():
-            # Number of agents in scenario.
-            agents_in_scenario = len(val.keys())
-            costs_list, counts_list = zip(
-                *[(record.costs, record.counts) for agent, record in val.items()]
-            )
-            # Sum costs over all agents in scenario.
-            costs_sum_agent: Costs = functools.reduce(
-                lambda a, b: add_dataclass(a, b), costs_list
-            )
-            # Average costs over number of agents in scenario.
-            costs_mean_agent = op_dataclass(costs_sum_agent, agents_in_scenario, divide)
-            # Sum costs over all scenarios.
-            costs_total = add_dataclass(costs_total, costs_mean_agent)
-            # Increment total number of episodes.
-            episodes += counts_list[0].episodes
-
-        # Average costs over total number of episodes.
-        costs_final = op_dataclass(costs_total, episodes, divide)
+        costs_final = avg_costs(records_sum=records_sum)
 
         # Compute sub-components of score.
         dist_to_destination = costs_final.dist_to_destination